Fix a deadlock problem with dispatch
Previously, when a worker's buffered requests channel filled up, a
deadlock was introduced. The worker is trying to signal that it is
finished with a piece of work on the `done` channel
(balancer.updateHeap). Meanwhile, the balancer routine is blocked
waiting on the worker's requests channel to open up space
(balancer.dispatch).

The fix required the balancer thread to be able to check whether or not
the requests channel is full without actually calling `requests <-
data`, which would always block. An exponential backoff retry function
is now used to delay the balancer in checking for a full request
channel. This allows the worker thread to complete some work, the
balancer to read it out of the `done` channel, thus unblocking the
balancer putting the next piece of data on the queue.