Fix balance tests
When the balancer's `Stop` function is called, the function will block
for two conditions: (1) the wait group for all worker threads to be
`Done` and (2) the balancer's `completed` channel to be full drained.
After both conditions are fulfilled, all data has been completely
handled, and it is safe to shutdown the balancer.